requirements.txt
langchain
langchain-community 
langchain-openai
langchain-experimental
python-dotenv
langchain_ollama
reportlab
rich
pdflab
PyMuPDF